chore: 컴포넌트와 스타일 코드 분리
