+++ src/components/Coupon/CouponItem/CouponItem.tsx
@@ -1,41 +1,39 @@
-import { css } from '@emotion/react';
 import { ChangeEvent } from 'react';
 
-import Checkbox from '../common/Checkbox';
+import {
+  couponDescription,
+  couponDescriptionWrapper,
+  couponItemContainer,
+  title,
+} from './CouponItem.styled';
 
-import { Coupon, CouponDiscountType } from '@/types/coupon';
+import Checkbox from '@/components/common/Checkbox/Checkbox';
+import { CouponClient } from '@/types/coupon';
 import { convertAvailableDateFormat, convertExpiryDateFormat } from '@/utils/date';
 
 interface CouponItemProps {
-  coupon: Coupon;
-  discountType: CouponDiscountType;
-  isCouponValid: boolean;
-  isChecked: boolean;
+  coupon: CouponClient;
+  isValid: boolean;
   handleChangeChecked: (e: ChangeEvent<HTMLInputElement>) => void;
 }
 
-const CouponItem = ({
-  coupon,
-  discountType,
-  isChecked,
-  isCouponValid,
-  handleChangeChecked,
-}: CouponItemProps) => {
-  const { description, expirationDate, minimumAmount, availableTime } = coupon;
+const CouponItem = ({ coupon, isValid, handleChangeChecked }: CouponItemProps) => {
+  const { id, isChecked, description, expirationDate, minimumAmount, availableTime, discountType } =
+    coupon;
 
   const isRequiredMinimumAmount =
     (discountType === 'fixed' || discountType === 'freeShipping') && minimumAmount;
   const isRequiredAvailableTime = discountType === 'percentage' && availableTime;
 
   return (
-    <div css={couponItemContainer(isCouponValid)}>
+    <div css={couponItemContainer(isValid)}>
       <Checkbox
         checked={isChecked}
         onChange={handleChangeChecked}
-        htmlFor={coupon.id.toString()}
+        htmlFor={id.toString()}
         label={description}
         labelCSS={title}
-        isDisabled={!isCouponValid}
+        isDisabled={!isValid}
       />
       <div css={couponDescriptionWrapper}>
         <span css={couponDescription}>만료일 : {convertExpiryDateFormat(expirationDate)}</span>
@@ -55,31 +53,3 @@ const CouponItem = ({
 };
 
 export default CouponItem;
-
-const couponItemContainer = (isCouponValid: boolean) => css`
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-
-  border-top: 1px solid #0000001a;
-  padding: 16px 0;
-
-  background-color: white;
-  opacity: ${isCouponValid ? 1 : 0.3};
-`;
-
-const title = css`
-  font-weight: 700;
-  font-size: 16px;
-`;
-
-const couponDescriptionWrapper = css`
-  display: flex;
-  flex-direction: column;
-  gap: 6px;
-`;
-
-const couponDescription = css`
-  font-weight: 500;
-  font-size: 12px;
-`;
