--- src/components/Coupon/CouponModal.tsx
@@ -1,103 +1,0 @@
-import { css } from '@emotion/react';
-import { Modal } from 'maru-nice-modal';
-import { useRecoilValue, useSetRecoilState } from 'recoil';
-
-import CouponItem from './CouponItem';
-import { isCheckedCoupon } from './utils/isCheckedCoupon';
-
-import { couponApplicabilityChecker } from '@/components/Coupon/utils/couponApplicabilityChecker';
-import { checkedCartItemsState, orderResultState } from '@/recoil/cartItems/selectors';
-import { couponSavedCheckListState } from '@/recoil/coupons/atoms';
-import { Coupon } from '@/types/coupon';
-import GuideText from '@common/GuideText';
-
-import useCoupon from '@hooks/useCoupon';
-
-interface CouponModalProps {
-  isOpen: boolean;
-  onClose: () => void;
-  couponList: Coupon[];
-}
-
-const CouponModal = ({ isOpen, onClose, couponList }: CouponModalProps) => {
-  const { couponCheckList, handleChangeChecked, isValidCouponCount, localDiscountPrice } =
-    useCoupon(couponList);
-  const { isCouponApplicable } = couponApplicabilityChecker(couponList);
-  const checkedCartItems = useRecoilValue(checkedCartItemsState);
-
-  const { totalOrderPrice } = useRecoilValue(orderResultState);
-  const setCouponSavedCheckList = useSetRecoilState(couponSavedCheckListState);
-
-  const isFreeShipping = isCheckedCoupon(couponCheckList, 'FREESHIPPING');
-  const displayDiscountLabel = `총 ${localDiscountPrice.toLocaleString('ko-KR')}원 할인 쿠폰 사용하기`;
-
-  const handleClickApplyCoupon = () => {
-    setCouponSavedCheckList(couponCheckList);
-    onClose();
-  };
-
-  return (
-    <Modal isOpen={isOpen}>
-      <Modal.Dimmed onDimmedClick={onClose} />
-      <Modal.Header>
-        <Modal.Title title="쿠폰을 선택해주세요" />
-        <Modal.CloseIcon onClose={onClose} />
-      </Modal.Header>
-      <Modal.Content css={contentWrapper}>
-        <GuideText label="쿠폰은 최대 2개까지 사용할 수 있습니다." />
-        <div css={couponListWrapper}>
-          {couponList.map((coupon, idx) => (
-            <CouponItem
-              key={coupon.code}
-              coupon={coupon}
-              discountType={coupon.discountType}
-              isCouponValid={
-                isValidCouponCount
-                  ? isCouponApplicable({
-                      coupon,
-                      totalOrderPrice: totalOrderPrice,
-                      checkedCartItems,
-                    })
-                  : couponCheckList[idx].isChecked
-              }
-              isChecked={couponCheckList[idx].isChecked}
-              handleChangeChecked={(e) => handleChangeChecked(e, coupon)}
-            />
-          ))}
-        </div>
-      </Modal.Content>
-      <Modal.ConfirmButton
-        css={confirmButton}
-        label={isFreeShipping ? `${displayDiscountLabel} + 무료 배송` : `${displayDiscountLabel}`}
-        onConfirm={handleClickApplyCoupon}
-      />
-    </Modal>
-  );
-};
-
-export default CouponModal;
-
-const contentWrapper = css`
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-
-  margin-bottom: 16px;
-`;
-
-const couponListWrapper = css`
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-`;
-
-const confirmButton = css`
-  width: 100%;
-  height: 44px;
-
-  border-radius: 5px;
-  background-color: #333333;
-
-  font-weight: 700;
-  color: white;
-`;
