+++ src/components/Coupon/CouponModal/CouponModal.styled.ts
@@ -1,0 +1,26 @@
+import { css } from '@emotion/react';
+
+export const contentWrapper = css`
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+
+  margin-bottom: 16px;
+`;
+
+export const couponListWrapper = css`
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+`;
+
+export const confirmButton = css`
+  width: 100%;
+  height: 44px;
+
+  border-radius: 5px;
+  background-color: #333333;
+
+  font-weight: 700;
+  color: white;
+`;
