+++ src/components/common/PriceSection/PriceSection.styled.ts
@@ -1,0 +1,16 @@
+import { css } from '@emotion/react';
+
+export const cartFooterSection = css`
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+`;
+
+export const borderTopWrapper = css`
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+
+  border-top: 1px solid #0000001a;
+  padding: 10px 0;
+`;
